+++ packages/pro/src/ai/llm.ts
@@ -4,7 +4,6 @@ import {
   AIProvider,
   ChatCompletionRequest,
   ConfigType,
-  LLMConfigOptions,
   LLMProviderConfig,
   Message,
   ResponseFormat,
@@ -13,32 +12,17 @@ import { tracer } from "dd-trace"
 import openai from "openai"
 import { z } from "zod"
 import { enrichAIConfig } from "../sdk/ai"
-import { Anthropic, AnthropicModel } from "./models"
-import { AzureOpenAI } from "./models/azureOpenai"
-import { LLM } from "./models/base"
-import { BudibaseAI } from "./models/budibaseai"
-import { OpenAI, OpenAIModel } from "./models/openai"
 
 // Provider-specific default models. If a provider's saved config does not
 // include a default model, we fall back to these. Azure defaults to gpt-4.1
 // so we avoid GPT-5-only parameters like verbosity/reasoning_effort.
-const DefaultModelByProvider: Record<AIProvider, OpenAIModel | AnthropicModel> =
-  {
-    OpenAI: "gpt-5-mini",
-    TogetherAI: "gpt-5-mini",
-    AzureOpenAI: "gpt-4.1",
-    Custom: "gpt-5-mini",
-    Anthropic: "claude-3-5-sonnet-20240620",
-    BudibaseAI: "gpt-5-mini",
-  }
-
-const ProviderMap = {
-  OpenAI: OpenAI,
-  TogetherAI: OpenAI,
-  AzureOpenAI: AzureOpenAI,
-  Custom: OpenAI,
-  Anthropic: Anthropic,
-  BudibaseAI: BudibaseAI,
+const DefaultModelByProvider: Record<AIProvider, string> = {
+  OpenAI: "gpt-5-mini",
+  TogetherAI: "gpt-5-mini",
+  AzureOpenAI: "gpt-4.1",
+  Custom: "gpt-5-mini",
+  Anthropic: "claude-3-5-sonnet-20240620",
+  BudibaseAI: "gpt-5-mini",
 }
 
 async function getAIConfig(): Promise<LLMProviderConfig | undefined> {
@@ -119,47 +103,6 @@ export async function getLLMConfig(): Promise<LLMProviderConfig | undefined> {
       // Next check for self-hosters that have their own API key.
       (await getSelfHostOpenAIKeyConfig())
   )
-}
-
-/**
- * @deprecated use the new `ai.sdk` instead
- */
-// This is the entrypoint for all LLM functionality in Budibase. If you're
-// making a feature that uses LLMs, you should call this function to get an LLM
-// instance. This function takes care of figuring out what LLM to use, and if
-// the user has no LLM configuration it will return undefined. It's the caller's
-// responsibility to handle this case.
-export async function getLLM(
-  options?: Omit<LLMConfigOptions, "model"> & { model?: string }
-): Promise<LLM | undefined> {
-  return await tracer.trace("getLLM", async span => {
-    const { model, maxTokens } = options || {}
-
-    const config = await getLLMConfig()
-    if (!config) {
-      span.addTags({ enabled: false, reason: "no config found" })
-      return
-    }
-
-    if (model) {
-      config.model = model
-    }
-    if (maxTokens) {
-      config.maxTokens = maxTokens
-    }
-
-    const LLMProvider = ProviderMap[config.provider]
-    if (!LLMProvider) {
-      span.addTags({
-        enabled: false,
-        reason: "no provider found",
-        provider: config.provider,
-      })
-      return
-    }
-
-    return new LLMProvider(config)
-  })
 }
 
 export class LLMRequest {
